Final enums corrections.
FInal corrections with enums.
