+++ MediMove/MediMove/Server/Entities/Transport.cs
@@ -3,29 +3,30 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using MediMove.Shared.Models.Enums;
 
 namespace MediMove.Shared.Entities
 {
-    public enum PatientPosition
-    {
-        Walking,
-        Sitting,
-        Lying
-
-    }
-
-    public enum Financing
-    {
-        FullyFunded,
-        PartiallyFunded,
-        FullyPaid
-    }
-
-    public enum TransportType
-    {
-        Visit,
-        Handover
-    }
+    //public enum PatientPosition
+    //{
+    //    Walking,
+    //    Sitting,
+    //    Lying
+    //
+    //}
+    //
+    //public enum Financing
+    //{
+    //    FullyFunded,
+    //    PartiallyFunded,
+    //    FullyPaid
+    //}
+    //
+    //public enum TransportType
+    //{
+    //    Visit,
+    //    Handover
+    //}
 
     public class Transport
     {
