+++ MediMove/MediMove/Shared/Models/DTOs/TransportDTO.cs
@@ -4,29 +4,30 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using MediMove.Shared.Models.Enums;
 
 namespace MediMove.Shared.Models.DTOs
 {
-    public enum PatientPosition
-    {
-        Walking,
-        Sitting,
-        Lying
-
-    }
-
-    public enum Financing
-    {
-        FullyFunded,
-        PartiallyFunded,
-        FullyPaid
-    }
-
-    public enum TransportType
-    {
-        Visit,
-        Handover
-    }
+    //public enum PatientPosition
+    //{
+    //    Walking,
+    //    Sitting,
+    //    Lying
+    //
+    //}
+    //
+    //public enum Financing
+    //{
+    //    FullyFunded,
+    //    PartiallyFunded,
+    //    FullyPaid
+    //}
+    //
+    //public enum TransportType
+    //{
+    //    Visit,
+    //    Handover
+    //}
 
     public class TransportsDTO
     {
@@ -41,6 +42,6 @@ namespace MediMove.Shared.Models.DTOs
         public PatientPosition PatientPosition { get; set; }
         public string Destination { get; set; }
         public TransportType TransportType { get; set; }
-
+        
     }
 }
